♻️ Modern C++17 Styled FileProtocol
- 모던 C++ 스타일 파일 프로토콜 재구축 완료

+++ CLIServer/ProtocolProvider.h
@@ -12,7 +12,7 @@ class CProtocolProvider
 {
 public:
 	static std::vector<unsigned char> GetPacket_CreateFile(std::string filepath);
-	static std::vector<unsigned char> GetPacket_WriteFile(int filesize);
+	static std::vector<unsigned char> GetPacket_WriteFile(unsigned long long filesize);
 	static std::vector<unsigned char> GetPacket_CloseFile();
 
 private:
